Move clean links to utils
