+++ src/nexgen/tools/vds_tools/__init__.py
@@ -1,6 +1,7 @@
 from nexgen.tools.vds_tools.utils import (
     VdsMapping,
     VdsSettings,
+    clean_unused_links,
     define_vds_dtype_from_bit_depth,
     find_datasets_in_file,
 )
@@ -10,4 +11,5 @@ __all__ = [
     "VdsSettings",
     "find_datasets_in_file",
     "define_vds_dtype_from_bit_depth",
+    "clean_unused_links",
 ]
